+++ test/integration/tests/activecredential.sh
@@ -8,7 +8,7 @@ cleanup() {
 
     # Evict persistent handles, we want them to always succeed and never trip
     # the onerror trap.
-    tpm2_evictcontrol -Q -a o -c 0x81010009 2>/dev/null || true
+    tpm2_evictcontrol -Q -C o -c 0x81010009 2>/dev/null || true
 
     if [ "$1" != "no-shut-down" ]; then
         shut_down