+++ test/integration/tests/createak.sh
@@ -7,8 +7,8 @@ cleanup() {
 
     # Evict persistent handles, we want them to always succeed and never trip
     # the onerror trap.
-    tpm2_evictcontrol -Q -a o -c 0x8101000b 2>/dev/null || true
-    tpm2_evictcontrol -Q -a o -c 0x8101000c 2>/dev/null || true
+    tpm2_evictcontrol -Q -C o -c 0x8101000b 2>/dev/null || true
+    tpm2_evictcontrol -Q -C o -c 0x8101000c 2>/dev/null || true
 
     # clear tpm state
     tpm2_clear
@@ -31,7 +31,7 @@ tpm2_createak -Q -C 0x8101000b -c ak.ctx -G rsa -g sha256 -s rsassa -u ak.pub -n
 tpm2_createak -C 0x8101000b -c ak.ctx -G rsa -g sha256 -s rsassa -u ak.pub -n ak.name
 tpm2_evictcontrol -c ak.ctx > ak.log
 phandle=`yaml_get_kv ak.log "persistent-handle"`
-tpm2_evictcontrol -Q -a o -c $phandle
+tpm2_evictcontrol -Q -C o -c $phandle
 
 # Test tpm2_createak with endorsement password
 cleanup "no-shut-down"