+++ test/integration/tests/load.sh
@@ -25,7 +25,7 @@ cleanup() {
 
   rm -f $file_load_key_pub $file_load_key_priv $file_load_key_name $file_load_key_ctx
 
-  tpm2_evictcontrol -Q -ao -c $Handle_parent 2>/dev/null || true
+  tpm2_evictcontrol -Q -Co -c $Handle_parent 2>/dev/null || true
 
   if [ $(ina "$@" "keep_ctx") -ne 0 ]; then
     rm -f $file_primary_key_ctx
@@ -53,7 +53,7 @@ tpm2_load -Q -C $file_primary_key_ctx  -u $file_load_key_pub  -r $file_load_key_
 
 cleanup "keep_ctx" "no-shut-down"
 
-tpm2_evictcontrol -Q -a o -c $file_primary_key_ctx -p $Handle_parent
+tpm2_evictcontrol -Q -C o -c $file_primary_key_ctx -p $Handle_parent
 
 tpm2_create -Q -C $Handle_parent   -g $alg_create_obj  -G $alg_create_key -u $file_load_key_pub  -r  $file_load_key_priv